Module tvm_graph_rt::packed_func

Enums

ArgValue
A borrowed TVMPODValue. Can be constructed using into() but the preferred way to obtain a ArgValue is automatically via call_packed!.

RetValue
An owned TVMPODValue. Can be converted from a variety of primitive and object types. Can be downcasted using try_from if it contains the desired type.

Traits

PackedFunc

Unions

TVMValue
\brief Union type of values being passed through API and function calls.