Crate tvm_graph_rt

This crate is an implementation of the TVM runtime for modules compiled with --system-lib. It’s mainly useful for compiling to WebAssembly and SGX, but also native if you prefer Rust to C++.

For TVM graphs, the entrypoint to this crate is runtime::GraphExecutor. Single-function modules are used via the packed_func! macro after obtaining the function from runtime::SystemLibModule

The main entrypoints to this crate are GraphExecutor For examples of use, please refer to the multi-file tests in the tests directory.

Re-exports

pub use self::errors::*;

Modules

errors

ffi

The low-level C runtime FFI API for TVM.

packed_func

Macros

call_packed

Calls a packed function and returns a RetValue.

import_module

Structs

DLTensor

\brief Plain C Tensor object, does not manage memory.

DsoModule

A module backed by a Dynamic Shared Object (dylib).

Entry

FuncCallError

Graph

A TVM computation graph.

GraphExecutor

A executor for a TVM computation graph.

Node

SystemLibModule

Tensor

A n-dimensional array type which can be converted to/from tvm::DLTensor and ndarray::Array. Tensor is primarily a holder of data which can be operated on via TVM (via DLTensor) or converted to ndarray::Array for non-TVM processing.

ValueDowncastError

Enums

ArgValue

A borrowed TVMPODValue. Can be constructed using into() but the preferred way to obtain a ArgValue is automatically via call_packed!.

RetValue

An owned TVMPODValue. Can be converted from a variety of primitive and object types. Can be downcasted using try_from if it contains the desired type.

Storage

A Storage is a container which holds Tensor data.

Constants

DTYPE_FLOAT32

DTYPE_FLOAT64

DTYPE_INT32

DTYPE_UINT32

Traits

Module

PackedFunc

Functions

TVMAPISetLastError⚠

TVMBackendAllocWorkspace

TVMBackendFreeWorkspace

TVMBackendParallelBarrier⚠

TVMBackendParallelLaunch

TVMGetLastError

load_param_dict

Loads a param dict saved using runtime.save_param_dict.

remove_item

Unions

TVMValue

\brief Union type of values being passed through API and function calls.