Struct tvm_graph_rt::Tensor

pub struct Tensor<'a> { /* private fields */ }

A n-dimensional array type which can be converted to/from tvm::DLTensor and ndarray::Array. Tensor is primarily a holder of data which can be operated on via TVM (via DLTensor) or converted to ndarray::Array for non-TVM processing.

Examples

extern crate ndarray;
use std::convert::TryInto;
use tvm_graph_rt::{call_packed, DLTensor, ArgValue, RetValue, Tensor};

let mut a_nd: ndarray::Array1<f32> = ndarray::Array::from_vec(vec![1f32, 2., 3., 4.]);
let mut a: Tensor = a_nd.into();
let mut a_dl: DLTensor = (&mut a).into();

let tvm_fn = |args: &[ArgValue]| -> Result<RetValue, ()> { Ok(RetValue::default()) };
call_packed!(tvm_fn, &mut a_dl);

// Array -> Tensor is mostly useful when post-processing TVM graph outputs.
let mut a_nd: ndarray::ArrayD<f32> = a.try_into().unwrap();

Implementations

impl<'a> Tensor<'a>

pub fn shape(&self) -> Vec<i64>

pub fn data(&self) -> &Storage<'_>

pub fn data_mut(&mut self) -> &'a mut Storage<'_>

pub fn to_vec<T: 'static + Debug + Clone>(&self) -> Vec<T>

Returns the data of this Tensor as a Vec.

Panics

Panics if the Tensor is not contiguous or does not contain elements of type T.

pub fn is_contiguous(&self) -> bool

Returns true iff this Tensor is represented by a contiguous region of memory.

pub fn copy(&mut self, other: &Tensor<'_>)

Returns a clone of this Tensor.

Panics

Panics if the Tensor is not contiguous or does not contain elements of type T.

pub fn to_owned(&self) -> Tensor<'static>

Returns an owned version of this Tensor via cloning.

pub fn as_dltensor(&self, flatten: bool) -> DLTensor